requirements.txt
apturl==0.5.2
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
einops==0.4.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.6.0
fonttools==4.29.1
future==0.18.2
gensim==4.1.2
h5py==3.6.0
httplib2==0.14.0
huggingface-hub==0.4.0
idna==2.8
imageio==2.16.1
joblib==1.1.0
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lmdb==1.3.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.1
monotonic==1.5
netifaces==0.10.4
networkx==2.7.1
numpy==1.22.2
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.19.4
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==5.3.1
regex==2022.3.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
sacremoses==0.0.47
scikit-image==0.19.2
scipy==1.8.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smart-open==5.2.1
ssh-import-id==5.10
systemd-python==234
tensorboardX==2.5
tifffile==2022.2.9
tokenizers==0.11.6
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.1.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
